Created ViewModel and Repository for PersonalInfo
Repository might be redundant in such simple case where all data comes from single source, but it should let to do any migration to another data sources in the future. As long as model would stay the same, ViewModel which has a lot of responsibilites will stay the same, and whole change will be in Repositories.
 Also in case of combining few data sources with some fallback it's moved out of cluttered ViewModel